requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.1.4
click==8.3.1
fastapi==0.132.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.2
packaging==26.0
pandas==3.0.1
psycopg==3.3.3
psycopg-binary==3.3.3
py4j==0.10.9.9
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyspark==3.5.8
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
rich==14.3.3
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.41.0
twelvedata
tvdatafeed
yahooquery